refactor: remove unused schedule manager
