+++ test/controllers/occurrences_manager_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.4.6 from annotations
-// in mona/test/controllers/schedule_occurrences_test.dart.
+// in mona/test/controllers/occurrences_manager_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
