--- test/controllers/schedule_manager_test.mocks.dart
@@ -1,51 +1,0 @@
-// Mocks generated by Mockito 5.4.6 from annotations
-// in mona/test/controllers/schedule_manager_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:mona/controllers/occurrences_manager.dart' as _i2;
-import 'package:mona/data/model/scheduled_occurrence.dart' as _i3;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: must_be_immutable
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-// ignore_for_file: invalid_use_of_internal_member
-
-/// A class which mocks [OccurrencesManager].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockOccurrencesManager extends _i1.Mock
-    implements _i2.OccurrencesManager {
-  @override
-  List<_i3.ScheduledOccurrence> current() => (super.noSuchMethod(
-        Invocation.method(
-          #current,
-          [],
-        ),
-        returnValue: <_i3.ScheduledOccurrence>[],
-        returnValueForMissingStub: <_i3.ScheduledOccurrence>[],
-      ) as List<_i3.ScheduledOccurrence>);
-
-  @override
-  List<_i3.ScheduledOccurrence> upcoming({required int? days}) =>
-      (super.noSuchMethod(
-        Invocation.method(
-          #upcoming,
-          [],
-          {#days: days},
-        ),
-        returnValue: <_i3.ScheduledOccurrence>[],
-        returnValueForMissingStub: <_i3.ScheduledOccurrence>[],
-      ) as List<_i3.ScheduledOccurrence>);
-}
